# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kaneo API",
    "description": "API specification for Kaneo",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:1337/api",
      "description": "API Server"
    }
  ],
  "paths": {
...
    "/public-project/{id}": {
      "get": {
        "operationId": "getPublic-project:id",
        "summary": "Get public project",
        "description": "Get a public project by ID (no authentication required)",
        "responses": {
          "200": {
            "description": "Public project details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "icon": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "description": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "isPublic": {
                      "type": "boolean"
                    },
                    "columns": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "position": {
                            "type": "number"
                          },
                          "tasks": {
                            "type": "array",
                            "items": {}
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "position",
                          "tasks"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "icon",
                    "slug",
                    "description",
                    "isPublic",
                    "columns"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Project is not public"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
...
  },
...
}